+++ src/App/CrashHandler/VRChatContentPublisher.CrashHandler.App/ViewModels/MainWindowViewModel.cs
@@ -1,0 +1,8 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+
+namespace VRChatContentPublisher.CrashHandler.App.ViewModels;
+
+public partial class MainWindowViewModel : ViewModelBase
+{
+    [ObservableProperty] public partial ViewModelBase? CurrentView { get; set; }
+}