+++ src/App/CrashHandler/VRChatContentPublisher.CrashHandler.App/ViewModels/ViewModelBase.cs
@@ -1,0 +1,7 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+
+namespace VRChatContentPublisher.CrashHandler.App.ViewModels;
+
+public abstract class ViewModelBase : ObservableObject
+{
+}