+++ src/App/CrashHandler/VRChatContentPublisher.CrashHandler.App/Views/CrashReportView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace VRChatContentPublisher.CrashHandler.App.Views;
+
+public partial class CrashReportView : UserControl
+{
+    public CrashReportView()
+    {
+        InitializeComponent();
+    }
+}